Worked on Encounter Scene
Completed setup for finding out which adjacent overworld cells make up
the arena. The game does not load adjacent sections when the player is
at the edge of a section in the overworld. The game does not load
adjacent sections when there is a "wall"  to the right and corners of
the player cell.